sdp_gateway/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
python-engineio==4.2.1
python-socketio==5.4.0
eventlet==0.33.0
pyOpenSSL==20.0.1
cryptography==3.4.7
python-dotenv==0.19.0 